+++ src/backend/storage/ipc/procsignal.c
@@ -416,8 +416,8 @@ WaitForProcSignalBarrier(uint64 generation)
 											5000,
 											WAIT_EVENT_PROC_SIGNAL_BARRIER))
 				ereport(LOG,
-						(errmsg("still waiting for backend with PID %lu to accept ProcSignalBarrier",
-								(unsigned long) slot->pss_pid)));
+						(errmsg("still waiting for backend with PID %d to accept ProcSignalBarrier",
+								(int) slot->pss_pid)));
 			oldval = pg_atomic_read_u64(&slot->pss_barrierGeneration);
 		}
 		ConditionVariableCancelSleep();
